+++ lib/src/tiled_map.dart
@@ -69,8 +69,17 @@ class TiledMap {
   List<Tileset> tilesets;
   List<Layer> layers;
 
-  String? backgroundColor;
-  Color? backgroundColor_;
+  /// Hex-formatted color (#RRGGBB or #AARRGGBB) to be rendered as a solid color
+  /// behind all other layers (optional).
+  String? backgroundColorHex;
+
+  /// [Color] to be rendered as a solid color behind all other layers
+  /// (optional).
+  ///
+  /// Parsed from [backgroundColorHex], will be null if parsing fails for any
+  /// reason.
+  Color? backgroundColor;
+
   int compressionLevel;
 
   int? nextLayerId;
@@ -97,8 +106,8 @@ class TiledMap {
     required this.tileHeight,
     this.tilesets = const [],
     this.layers = const [],
+    this.backgroundColorHex,
     this.backgroundColor,
-    this.backgroundColor_,
     this.compressionLevel = -1,
     this.hexSideLength,
     this.nextLayerId,
@@ -226,8 +235,8 @@ class TiledMap {
   }
 
   static TiledMap parse(Parser parser, {List<TsxProvider>? tsxList}) {
-    final backgroundColor = parser.getStringOrNull('backgroundcolor');
-    final backgroundColor_ = parser.getColorOrNull('backgroundcolor');
+    final backgroundColorHex = parser.getStringOrNull('backgroundcolor');
+    final backgroundColor = parser.getColorOrNull('backgroundcolor');
     final compressionLevel = parser.getInt('compressionlevel', defaults: -1);
     final height = parser.getInt('height');
     final hexSideLength = parser.getIntOrNull('hexsidelength');
@@ -283,8 +292,8 @@ class TiledMap {
       tileHeight: tileHeight,
       tilesets: tilesets,
       layers: layers,
+      backgroundColorHex: backgroundColorHex,
       backgroundColor: backgroundColor,
-      backgroundColor_: backgroundColor_,
       compressionLevel: compressionLevel,
       hexSideLength: hexSideLength,
       nextLayerId: nextLayerId,