+++ test/object_group_test.dart
@@ -28,10 +28,10 @@ void main() {
     });
 
     test('sets color', () {
-      expect(objectGroup.color, equals('#555500'));
+      expect(objectGroup.colorHex, equals('#555500'));
       expect(
-        objectGroup.color_,
-        equals(const Color.fromARGB(255, 85, 85, 0)),
+        objectGroup.color,
+        equals(Color(int.parse('0xff555500'))),
       );
     });