+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## 0.8.5
+## [Next]
 * Add parser support for Tiled hex colors as `ui.Color`
+* BREAKING CHANGE: re-named string color fields to `colorHex` such as `layer.tintColorHex` and 
+  added a new `color` field paired with each `colorHex` field of type `ui.Color`.
 
 ## 0.8.4
 * Adding "class" attribute for Tiled 1.9's Unified Custom Types: